requirements.txt
h5py
transformers
numpy
Pillow
torchvision
tqdm